Fix simplest ctor to avoid null ref
If the simpler constructor was used, the "lockOperation" member was not
being initialized and was triggering null ref later.  That being said,
the lockOperation member itself is unnecessary as the LockOperation
method satisfies Func<bool> directly.  Also updated the simple ctor to
call the more complex one with null for acl to reduce likelihood of this
kind of bug being introduced in the future.