Support dns round-robin in connection string
Based on "http://wiki.apache.org/hadoop/ZooKeeper/FAQ#A8", I originally
tried leveraging dns round-robin, but it turns out that ZooKeeperDotNet
didn't support this.  Here is a change that will permit the client to be
initialized with a "connection string" that contains hostnames or ip
addresses, and will resolve a hostname to all associated IP addresses.