All projects: Update to .NET Standard 2.0
This is a refresh of PR #37, initially submitted by Maciej Chyczewski:

    https://github.com/ewhauser/zookeeper/pull/37#issuecomment-615344633

The project files have been rebuilt using .NET Core 2, and reference
their 'log4net' and 'NUnit' dependencies via NuGet packages.

The result was built and tested on the following implementations:

  * .NET Core 2.2.401, NixOS
  * .NET Core 3.1.101, Windows 10
  * Visual Studio Professional 2017, Version 15.9.7
  * Visual Studio Community 2019, Version 16.4.4

The assemblies can be consumed from .NET 4.7.2 (tested) and other
.NET-Standard-2.0-aware runtimes.

This commit also includes a refreshed S22.Sasl, which is used by the
SASL tests.  It was built by running 'dotnet build' on this commit of
https://github.com/ztzg/S22.Sasl:

    commit 85c89c6acb9e42e7d9cc4208e2e0c63aebde1e69
    Author: Damien Diederen <dd@crosstwine.com>
    Date:   Mon Apr 27 12:25:59 2020 +0200

        S22.Sasl: Migrate to .NET Standard 2.0

which is available on this branch:

    https://github.com/ztzg/S22.Sasl/tree/RT-46545-zookeeper-net-sasl-binary-netstandard2.0

Co-authored-by: Maciej Chyczewski <maciej@rubrikkgroup.com>