ZOOKEEPER-4030: Remove WrapperInetSocketAddress and WrapperInetAddress
On JDK 8+, InetAddress's methods are not final; there is consequently
no need to wrap that class.

After that, the only remaining purpose of WrapperInetSocketAddress is
to allow mocking InetSocketAddress.getAddress().  We can as well pass
the address to a testing-visible only entry point as there is not much
value in testing getAddress per se.