ZOOKEEPER-2454: Add infrastructure to keep track of authenticated connections
This change only adds supporting infrastructure, and does not
implement any policy.

It started as an (accidental) parallel implementation of the ideas
presented by Botond Hejj in his ZOOKEEPER-2454 patch dated 2016-06-27.
The features and vocabularly of that patch have later been adopted;
this is the "merged" result.

An important difference with the previous patch is that the policy is
delegated to a "pluggable" interface.  (Hardcoded to 'null' here; the
configuration mechanism is to be added in a subsequent patch.)

It also supports Netty in addition to NIO.

Co-authored-by: Botond Hejj <botond.hejj@gmail.com>