ZOOKEEPER-2454: Fix: Process AUTH_FAILED during SASL tunnel auth
Without this, the Java client suffers an exception and fails to set
its state to AUTH_FAILED if an error is signaled via the 'err' field
of the reply.

The exception is relatively harmless, but the state not being set
causes the client to try and reconnect, unlike what happens with
SASL-level (or OpCode.auth) errors.