ZOOKEEPER-2454: Continue returning AUTHFAILED to OpCode.auth clients
ZooKeeper would invariably reply to clients with AUTHFAILED if the
authentication provider failed to return OK.

A previous change (Interrogate authentication limiter for Opcode.auth
requests) was careful to preserve and propagate the actual error code
returned by the provider.

Unfortunately, that change is not "bug-for-bug backward compatible" as
AUTHFAILED is handled specially by clients--whereas deployed providers
may be "accidentally" returning other codes, not realizing that that
causes clients to reconnect if the code is actually propagated.

This patch partially reverts to the old behavior (at least when it
comes to providers), overriding potentially problematic codes with
AUTHFAILED (but logging a warning rather than doing so silently).